+++ CHANGELOG.md
@@ -11,6 +11,8 @@ the code was deployed.
 
 ## [Unreleased]
 
+## [1.3.0] - 2022-07-26
+
 ### Changed
 
 - Improved error messaging in Twilio Number Purchasing.
@@ -45,7 +47,8 @@ the code was deployed.
 
 - Initial version (CU-21ghk0x).
 
-[unreleased]: https://github.com/bravetechnologycoop/particle-accelerator/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/bravetechnologycoop/particle-accelerator/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/bravetechnologycoop/particle-accelerator/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/bravetechnologycoop/particle-accelerator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/bravetechnologycoop/particle-accelerator/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/bravetechnologycoop/BraveSensor/releases/tag/v1.0.0